requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
distlib==0.3.8
Django==5.1
django-cors-headers==3.11.0
django-extensions==3.2.3
filelock==3.15.4
firebase==4.0.1
firebase-admin==6.5.0
future==1.0.0
gcloud==0.18.3
gevent==24.2.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.65.5
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
msgpack==1.0.8
oauth2client==4.1.3
pipenv==2024.0.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.2
Pyrebase4==4.8.0
PySocks==1.7.1
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.26.3
zope.event==5.0
zope.interface==7.0.1
